Move history to slices
