+++ src/Slices/ServiceInstanceHistory/index.ts
@@ -1,0 +1,1 @@
+export * as ServiceInstanceHistory from "./Core/Route";
