+++ src/Test/Data/index.ts
@@ -1,5 +1,4 @@
 export * as Diagnose from "./Diagnose";
-export * as InstanceLog from "./InstanceLog";
 export * as Field from "./Field";
 export * as Row from "./Row";
 export * as Service from "./Service";
